+++ src/Tes.Runner/Exceptions/FatalException.cs
@@ -1,0 +1,9 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Tes.Runner.Exceptions
+{
+    public abstract class FatalException : Exception
+    {
+    }
+}
