+++ src/Tes.Runner/Exceptions/IdentityUnavailableFatalException.cs
@@ -1,15 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace Tes.Runner.Authentication
+namespace Tes.Runner.Exceptions
 {
-    public abstract class FatalException : Exception
-    {
-
-    }
-
     public class IdentityUnavailableException : FatalException
     {
-
     }
 }
